Made the flake more useful.
Updated the flake to install actual dependencies and removed conan from
the flake packages. Conan packages was unable to build on nixos due to
the posix file standard. So I installed the build dependencies directly
in the flake instead.

Also updated the (c)makefiles so that they should still work with conan
if someone wants to build with that.